+++ app/apps/task/adapter/task_adapter.go
@@ -23,7 +23,7 @@ type TaskDetailInput struct {
 }
 
 type TaskDto struct {
-	ID          int64     `json:"id"`
+	Id          int64     `json:"id"`
 	UserID      int64     `json:"user_id"`
 	ProjectID   int64     `json:"project_id"`
 	StatusID    int64     `json:"status_id"`
@@ -42,7 +42,7 @@ type TaskDto struct {
 
 func ToTaskDto(task *domain.Task) *TaskDto {
 	out := &TaskDto{
-		ID:        task.ID,
+		Id:        task.Id,
 		UserID:    task.UserID,
 		ProjectID: task.ProjectID,
 		StatusID:  task.StatusID,