+++ app/domain/project.go
@@ -4,27 +4,39 @@ import (
 	"context"
 	"time"
 
-	"gitlab.com/shaninalex/lumna/app/internal/db"
-	"gitlab.com/shaninalex/lumna/app/internal/utils"
+	"gitlab.com/shaninalex/lumna/app/pkg/db"
+	"gitlab.com/shaninalex/lumna/app/pkg/utils"
 )
 
 type Project struct {
-	ID        int64
+	Id        int64
 	Title     string
 	Code      string
+	UserId    int64
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
 
-func MakeProject(project *db.Project) *Project {
-	return &Project{
-		ID:        project.ID,
-		Title:     project.Title,
-		Code:      project.Code,
-		CreatedAt: project.CreatedAt,
-		UpdatedAt: project.UpdatedAt,
-	}
-}
+// GetID - returns the id.
+func (s *Project) GetID() int64 { return s.Id }
+
+// SetID - sets the id.
+func (s *Project) SetID(id int64) { s.Id = id }
+
+// GetOwnerID - returns the owner id.
+func (s *Project) GetOwnerID() int64 { return s.UserId }
+
+// IsOwner - checks if it is owner.
+func (s *Project) IsOwner(entity AuthUser) bool { return entity.GetID() == s.GetOwnerID() }
+
+// GetCreatedAt - returns the created at.
+func (s *Project) GetCreatedAt() time.Time { return s.CreatedAt }
+
+// GetUpdatedAt - returns the updated at.
+func (s *Project) GetUpdatedAt() time.Time { return s.UpdatedAt }
+
+// GetCreatedBy - returns the created by.
+func (s *Project) GetCreatedBy() int64 { return s.GetOwnerID() }
 
 // ProjectReader - project reader.
 type ProjectReader interface {
@@ -53,13 +65,13 @@ func NewProjectService() *ProjectService {
 
 func (p ProjectService) List(ctx context.Context) ([]*Project, error) {
 	connection := db.GetDb(ctx)
-	projects, err := db.ProjectList(ctx, connection)
+	projects, err := ProjectList(ctx, connection)
 	if err != nil {
 		return nil, err
 	}
 	output := make([]*Project, len(projects))
 	for i, project := range projects {
-		output[i] = MakeProject(project)
+		output[i] = project
 	}
 
 	return output, nil
@@ -67,24 +79,24 @@ func (p ProjectService) List(ctx context.Context) ([]*Project, error) {
 
 func (p ProjectService) GetProject(ctx context.Context, id int64) (*Project, error) {
 	connection := db.GetDb(ctx)
-	project, err := db.ProjectGetByID(ctx, connection, id)
+	project, err := ProjectGetByID(ctx, connection, id)
 	if err != nil {
 		return nil, err
 	}
-	return MakeProject(project), nil
+	return project, nil
 }
 
 func (p ProjectService) CreateProject(ctx context.Context, project *Project) (*Project, error) {
-	_project := &db.Project{
+	_project := &Project{
 		Title: project.Title,
 		Code:  utils.GenerateEntityCode("project"),
 	}
-	err := db.ProjectSave(ctx, db.GetDb(ctx), _project)
+	err := ProjectSave(ctx, db.GetDb(ctx), _project)
 	if err != nil {
 		return nil, err
 	}
 	now := time.Now()
-	project.ID = _project.ID
+	project.Id = _project.Id
 	project.Code = _project.Code
 	project.CreatedAt = now
 	project.UpdatedAt = now
@@ -92,11 +104,11 @@ func (p ProjectService) CreateProject(ctx context.Context, project *Project) (*P
 }
 
 func (p ProjectService) UpdateProject(ctx context.Context, project *Project) (*Project, error) {
-	err := db.ProjectUpdate(ctx, db.GetDb(ctx), &db.Project{ID: project.ID, Title: project.Title})
+	err := ProjectUpdate(ctx, db.GetDb(ctx), &Project{Id: project.Id, Title: project.Title})
 	if err != nil {
 		return nil, err
 	}
-	project, err = p.GetProject(ctx, project.ID)
+	project, err = p.GetProject(ctx, project.Id)
 	if err != nil {
 		return nil, err
 	}
@@ -104,5 +116,5 @@ func (p ProjectService) UpdateProject(ctx context.Context, project *Project) (*P
 }
 
 func (p ProjectService) DeleteProject(ctx context.Context, id int64) error {
-	return db.ProjectDelete(ctx, db.GetDb(ctx), id)
+	return ProjectDelete(ctx, db.GetDb(ctx), id)
 }