+++ app/domain/user.go
@@ -2,9 +2,11 @@ package domain
 
 import (
 	"context"
+	"encoding/json"
+	"log"
 	"time"
 
-	"gitlab.com/shaninalex/lumna/app/internal/db"
+	"gitlab.com/shaninalex/lumna/app/pkg/db"
 	"golang.org/x/crypto/bcrypt"
 )
 
@@ -15,17 +17,57 @@ type (
 
 	// User user model
 	User struct {
-		ID       int64
-		Email    string
-		Settings string
-		Active   bool
-		State    UserState
-		Code     string
+		Id           int64
+		Email        string
+		Settings     string
+		Active       bool
+		State        UserState
+		Code         string
+		PasswordHash string `json:"-"`
 
 		CreatedAt time.Time
 		UpdatedAt time.Time
 	}
 )
+
+// GetID - returns the id.
+func (s *User) GetID() int64 { return s.Id }
+
+// SetID - sets the id.
+func (s *User) SetID(id int64) { s.Id = id }
+
+// IsActive - checks if it is active.
+func (s *User) IsActive() bool { return s.Active }
+
+// GetCreatedAt - returns the created at.
+func (s *User) GetCreatedAt() time.Time { return s.CreatedAt }
+
+// GetUpdatedAt - returns the updated at.
+func (s *User) GetUpdatedAt() time.Time { return s.UpdatedAt }
+
+// SetCode - sets the code.
+func (s *User) SetCode(code string) { s.Code = code }
+
+// GetCode - returns the code.
+func (s *User) GetCode() string { return s.Code }
+
+func (s *User) GetSettings() *UserSettings {
+	var settings UserSettings
+	err := json.Unmarshal([]byte(s.Settings), &settings)
+	if err != nil {
+		log.Println("User.GetSettings. Error:", err)
+		return &DefaultUserSettings
+	}
+	return &settings
+}
+
+func (s *User) SetSettings(settings *UserSettings) {
+	b, err := json.Marshal(&settings)
+	if err != nil {
+		panic(err)
+	}
+	s.Settings = string(b)
+}
 
 const (
 	UserStatePending UserState = "pending"
@@ -35,9 +77,9 @@ const (
 )
 
 type UserManager interface {
-	GetUser(ctx context.Context, userID int64) (*db.User, error)
-	GetUserByEmail(ctx context.Context, email string) (*db.User, error)
-	UpdateUserSettings(ctx context.Context, userID int64, settings *db.UserSettings) error
+	GetUser(ctx context.Context, userID int64) (*User, error)
+	GetUserByEmail(ctx context.Context, email string) (*User, error)
+	UpdateUserSettings(ctx context.Context, userID int64, settings *UserSettings) error
 	SetPassword(ctx context.Context, userId int64, rawPwd string) error
 	CheckPassword(ctx context.Context, userId int64, rawPassword string) error
 	CreateUser(ctx context.Context, email string, rawPassword string) (*User, error)
@@ -52,8 +94,8 @@ func NewUserService() *UserService {
 }
 
 // GetUser get user
-func (s UserService) GetUser(ctx context.Context, userId int64) (*db.User, error) {
-	user, err := db.UserGetByField(ctx, db.GetDb(ctx), "id", userId)
+func (s UserService) GetUser(ctx context.Context, userId int64) (*User, error) {
+	user, err := UserGetByField(ctx, db.GetDb(ctx), "id", userId)
 	if err != nil {
 		return nil, err
 	}
@@ -61,8 +103,8 @@ func (s UserService) GetUser(ctx context.Context, userId int64) (*db.User, error
 }
 
 // GetUserByEmail get user by email
-func (s UserService) GetUserByEmail(ctx context.Context, email string) (*db.User, error) {
-	user, err := db.UserGetByField(ctx, db.GetDb(ctx), "email", email)
+func (s UserService) GetUserByEmail(ctx context.Context, email string) (*User, error) {
+	user, err := UserGetByField(ctx, db.GetDb(ctx), "email", email)
 	if err != nil {
 		return nil, err
 	}
@@ -70,15 +112,15 @@ func (s UserService) GetUserByEmail(ctx context.Context, email string) (*db.User
 }
 
 // UpdateUserSettings update user settings
-func (s UserService) UpdateUserSettings(ctx context.Context, userId int64, settings *db.UserSettings) error {
+func (s UserService) UpdateUserSettings(ctx context.Context, userId int64, settings *UserSettings) error {
 	connection := db.GetDb(ctx)
-	user, err := db.UserGetByField(ctx, connection, "id", userId)
+	user, err := UserGetByField(ctx, connection, "id", userId)
 	if err != nil {
 		return err
 	}
 	user.SetSettings(settings)
 	user.UpdatedAt = time.Now()
-	if err = db.UserUpdate(ctx, connection, user); err != nil {
+	if err = UserUpdate(ctx, connection, user); err != nil {
 		return err
 	}
 	return nil
@@ -91,7 +133,7 @@ func (s UserService) SetPassword(ctx context.Context, userId int64, rawPwd strin
 
 // CheckPassword check password
 func (s UserService) CheckPassword(ctx context.Context, userId int64, password string) error {
-	user, err := db.UserGetByField(ctx, db.GetDb(ctx), "id", userId)
+	user, err := UserGetByField(ctx, db.GetDb(ctx), "id", userId)
 	if err != nil {
 		return err
 	}
@@ -109,19 +151,19 @@ func (s UserService) CreateUser(ctx context.Context, email, rawPassword string) 
 		return nil, err
 	}
 
-	user := &db.User{
+	user := &User{
 		Email:        email,
 		PasswordHash: string(hash),
 	}
 
 	connection := db.GetDb(ctx)
-	user, err = db.UserSave(ctx, connection, user)
+	user, err = UserSave(ctx, connection, user)
 	if err != nil {
 		return nil, err
 	}
 
 	return &User{
-		ID:        user.ID,
+		Id:        user.Id,
 		Email:     user.Email,
 		Settings:  user.Settings,
 		Active:    user.Active,